python_requirements.txt
llama-index
python-dotenv
httpx
beautifulsoup4
llama-index-llms-openai-like
